==> 1.data_gathering.ipynb <==
import pandas as pd
from utils.api_keys import fetch_api_key
from utils.downloader import TMDBMovieDownloader

downloader = TMDBMovieDownloader(fetch_api_key("tmdb", False),
                            filepath="data/movie_data.csv",
                            filepath_creds="data/movie_credits.csv",
                            batch_size=50,
                            max_batches=10)

downloader.download_data()

df = pd.read_csv('data/movie_data.csv')

df

df.shape

df.columns

df.duplicated().sum()

df.isnull().sum()
