==> 2.data_cleaning.ipynb <==
import os
import pandas as pd
import json
from datetime import datetime

movies = pd.read_csv('data/movie_data.csv')
creds = pd.read_csv("data/movie_credits.csv")

movies.shape, creds.shape

movies.head()

movies.columns

movies.isnull().sum()

movies[movies['budget']==0].shape

movies.columns

metadata_cols = ['adult', 'backdrop_path', 'belongs_to_collection', 'budget', 'homepage', 'imdb_id', 'popularity', 'poster_path', 'production_countries', 'revenue', 'runtime', 'status', 'video', 'vote_average', 'vote_count', 'spoken_languages', 'original_title']

metadata = movies[metadata_cols+['id']]
movies = movies.drop(columns=metadata_cols)

movies

def process_list_dict(entry, jload=True):
    if isinstance(entry, float):
        return entry
    entry = json.loads(entry) if jload else entry
    res = []
    for d in entry:
        res.append(d['name'])
    return ", ".join(res)
def process_list(entry, jload=True):
    if isinstance(entry, float):
        return entry
    entry = json.loads(entry) if jload else entry
    return ", ".join(entry)
def process_dict_list_dict(entry, jload=True):
    if isinstance(entry, float):
        return entry
    entry = json.loads(entry)['keywords'] if jload else entry
    return process_list_dict(entry, False)
def process_date(entry):
    if isinstance(entry, float):
        return entry
    return datetime.strptime(entry, "%Y-%m-%d")
def process_dict(entry, jload=True):
    if isinstance(entry, float):
        return entry
    entry = json.loads(entry) if jload else entry
    return entry['name']

movies['keywords'] = movies['keywords'].apply(process_dict_list_dict)
movies['production_companies'] = movies['production_companies'].apply(process_list_dict)
movies['origin_country'] = movies['origin_country'].apply(process_list)
movies['genres'] = movies['genres'].apply(process_list_dict)
movies['release_date'] = movies['release_date'].apply(process_date)

movies['release_year'] = movies['release_date'].dt.year
metadata['release_date'] = movies['release_date']
movies = movies.drop(columns=['release_date'])

movies.head()

metadata['production_countries'] = metadata['production_countries'].apply(process_list_dict)
metadata['belongs_to_collection'] = metadata['belongs_to_collection'].apply(process_dict)
metadata['spoken_languages'] = metadata['spoken_languages'].apply(process_list_dict)

metadata.head()

movies.head()

missing = "[MISSING]"
print(movies.isnull().sum())
movies = movies.fillna(value=missing)

def get_na(df):
    res = {}
    for col in df.columns:
        if isinstance(df[col][0], str):
            res[col] = len(df[df[col].apply(len)==0])
    return res

print(get_na(movies))

def set_na(df):
    for col in df.columns:
        if isinstance(df[col][0], str):
            df[col] = df[col].apply(lambda x: missing if len(x)==0 else x)

set_na(movies)

movies.head()

metadata.head()

metadata.isnull().sum()

metadata = metadata.fillna(value=missing)

get_na(metadata)

set_na(metadata)

metadata

creds

def extract_actors(cast, jload = True):
    cast = json.loads(cast) if jload else cast
    res = []
    for d in cast:
        if d['known_for_department'] == "Acting":
            res.append(d['name'])
    res = res[:10]
    return ", ".join(res).strip()
def extract_director(crew, jload = True):
    crew = json.loads(crew) if jload else crew
    res = []
    for d in crew:
        if d['job'] == 'Director':
            res.append(d['name'])
    res = res[:10]
    return ", ".join(res).strip()
def extract_writer(crew, jload = True):
    crew = json.loads(crew) if jload else crew
    res = []
    for d in crew:
        if d['job'] == 'Writer':
            res.append(d['name'])
    res = res[:10]
    return ", ".join(res).strip()
def get_crew(entry, jload = True):
    entry = json.loads(entry) if jload else entry
    res = []
    for d in entry:
        res.append(f"{d['job']}: {d['name']}")
    res.sort()
    return ", ".join(res)

creds['writers'] = creds['crew'].apply(extract_writer)
creds['directors'] = creds['crew'].apply(extract_director)
creds['actors'] = creds['cast'].apply(extract_actors)
creds['cast'] = creds['cast'].apply(process_list_dict)
creds['crew'] = creds['crew'].apply(get_crew)

set_na(creds)

creds

metadata = pd.merge(metadata, creds[['movie_id', 'cast', 'crew']], left_on='id', right_on='movie_id', how='outer')

metadata.head()

creds = creds.drop(columns=['crew', 'cast'])
creds.head()

movies = pd.merge(movies, creds.drop(columns=['movie_title']), left_on='id', right_on='movie_id', how='outer' ).drop(columns='movie_id')

movies.head()

movies.columns

movies.head()

os.makedirs("data/cleaned-1", exist_ok=True)
movies.to_csv(path_or_buf= "data/cleaned-1/movies.csv")

metadata = pd.merge(metadata.drop(columns=['backdrop_path', 'production_countries', 'video', 'movie_id', 'adult', "status"]),
         movies.drop(columns=['writers', 'directors', 'actors', 'release_year']), on='id', how='outer' )

metadata.to_csv(path_or_buf= "data/cleaned-1/metadata.csv")
