==> 3.data_preprocessing.ipynb <==
import os
import pandas as pd

movies = pd.read_csv('data/cleaned-1/movies.csv').drop(columns='Unnamed: 0')
metadata = pd.read_csv('data/cleaned-1/metadata.csv').drop(columns='Unnamed: 0')

movies

def convert(entry):
    elems = entry.split(", ")
    res = []
    for elem in elems:
        res.append(elem.replace(" ", ""))
    return " ".join(res)
def convert2(entry):
    return entry.replace(",", "")

movies.columns

movies['keywords'] = movies['keywords'].apply(convert2)
movies['genres']  = movies['genres'].apply(convert2)

movies['production_companies'] = movies['production_companies'].apply(convert)
movies['writers'] = movies['writers'].apply(convert)
movies['directors'] = movies['directors'].apply(convert)
movies['actors'] = movies['actors'].apply(convert)

movies

movies

def get_soup(row):
    sep = ". "
    missing = "[MISSING]"
    res = (f"Title: {row['title']}" + sep) if row['title']!=missing else ""
    res += (f"Genres: {row['genres']}" + sep) if row['genres']!=missing else ""
    res += (f"Keywords: {row['keywords']}" + sep) if row['keywords']!=missing else ""
    res += (f"Cast: {row['actors']}" + sep) if row['actors']!=missing else ""
    res += (f"Directors: {row['directors']}" + sep) if row['directors']!=missing else ""
    res += (f"Writers: {row['writers']}" + sep) if row['writers']!=missing else ""
    res += (f"Overview: {row['overview']}" + sep) if row['overview']!=missing else ""
    res += (f"Tagline: {row['tagline']}" + sep) if row['tagline']!=missing else ""
    res += (f"Released in: {row['release_year']}" + sep) if row['release_year']!=missing else ""
    res += (f"Productions: {row['production_companies']}" + sep) if row['production_companies']!=missing else ""
    res += (f"Original Language: {row['original_language']}" + sep) if row['original_language']!=missing else ""
    return res[:-len(sep)]


movies['soup'] = movies.apply(get_soup, axis=1)

movies

metadata

os.makedirs("data/cleaned-2", exist_ok=True)
movies.to_csv(path_or_buf= "data/cleaned-2/movies.csv")
metadata.to_csv(path_or_buf= "data/cleaned-2/metadata.csv")
