==> 3.modeling.ipynb <==
import os

import pandas as pd

movies = pd.read_csv('data/cleaned-2/movies.csv').drop(columns='Unnamed: 0')
metadata = pd.read_csv('data/cleaned-2/metadata.csv').drop(columns='Unnamed: 0')

from langchain.docstore.document import Document

metadata['soup'] = movies['soup']

metadata.columns

movies = []

for idx, row in metadata.iterrows():
    x = Document(page_content=row['soup'],
                 metadata = {
                     "movie": row['title'],
                     "tmdb_id": row['id'],
                     "imdb_id": row['imdb_id'],
                     'genres': row['genres'],
                     "release_date": row['release_date'],
                     "cast": row['cast'],
                     "crew": row['crew'],
                     "collection": row['belongs_to_collection'],
                     "budget": row['budget'],
                     "revenue": row['revenue'],
                     "runtime": row['runtime'],
                     "language": row['original_language'],
                     "popularity": row['popularity'],
                     "synopsis": row['overview'],
                     "poster_path": row['poster_path'],
                     "homepage": row['homepage']
                 })
    movies.append(x)

movies[0]

len(movies)

import weaviate
from utils.api_keys import fetch_api_key, add_api_key

add_api_key("weaviate2")

# client = weaviate.Client(
#     url="https://cjscqn5qa6ddwpdf5ucna.c0.asia-southeast1.gcp.weaviate.cloud",
#     auth_client_secret=weaviate.AuthApiKey(fetch_api_key("weaviate", False))
# )
client = weaviate.Client(
    url="https://vz10aml8qwbzfzxohgk7a.c0.asia-southeast1.gcp.weaviate.cloud",
    auth_client_secret=weaviate.AuthApiKey(fetch_api_key("weaviate2", False)))

from langchain_huggingface import HuggingFaceEmbeddings

embedding_model_name = "sentence-transformers/all-MiniLM-L6-v2"
model_kwargs = {"device": "cuda"}
embeddings = HuggingFaceEmbeddings(
  model_name=embedding_model_name,
  model_kwargs=model_kwargs
)

from langchain.vectorstores import Weaviate

# Weaviate()

# vector_db = Weaviate.from_documents(
#     movies,
#     embeddings,
#     client = client,
#     by_text = False
# )
vector_db = Weaviate(client=client,embedding=embeddings,
                      index_name="Mrsprj",
                      text_key="movies",
                      by_text=False,
                     attributes=list(movies[0].metadata.keys()),
                     )

# vector_db.add_documents(movies)

soups = pd.Series(metadata['soup'].values, index=metadata['title'])

soups

def get_recommendations(title, vector_db):

    if title not in soups:
        raise ValueError(f"Title '{title}' not found in indices")

    query = soups.get(title)

    try:
        results = vector_db.similarity_search_with_score(query, k=11)

        top_ten = []

        for x in results[1:]:

            movie_metadata = {
                'movie': x[0].metadata['movie'],
                "tmdb_id": x[0].metadata['tmdb_id'],
                "imdb_id": x[0].metadata['imdb_id'],
                'genres': x[0].metadata['genres'],
                "release_date": x[0].metadata['release_date'],
                "cast": x[0].metadata['cast'],
                "crew": x[0].metadata['crew'],
                "collection": x[0].metadata['collection'],
                "budget": x[0].metadata['budget'],
                "revenue": x[0].metadata['revenue'],
                "runtime": x[0].metadata['runtime'],
                "language": x[0].metadata['language'],
                "popularity": x[0].metadata['popularity'],
                "synopsis": x[0].metadata['synopsis'],
                "poster_path": x[0].metadata['poster_path'],
                "homepage": x[0].metadata['homepage'],
                'similarity_score': round(x[1], 2),
            }
            top_ten.append(movie_metadata)

        df_top_ten = pd.DataFrame(top_ten)
        return df_top_ten

    except Exception as e:
        print(f"Error during query: {e}")
        return None

metadata.title

querry = metadata[metadata['title'].str.contains('Star Wars')]['title'].values[0]
rec = get_recommendations(querry, vector_db)

rec

rec.columns

metadata[metadata['title'].str.contains("Vampire")]['title'].values[0]

metadata['id']

metadata['soup']

metadata

def get_recommendations(title, vector_db):

    if title not in soups:
        raise ValueError(f"Title '{title}' not found in indices")

    query = soups.get(title)

    try:
        results = vector_db.similarity_search_with_score(query, k=11)

        top_ten = []

        for x in results[1:]:

            movie_metadata = {
                'movie': x[0].metadata['movie'],
                "tmdb_id": x[0].metadata['tmdb_id'],
                "imdb_id": x[0].metadata['imdb_id'],
                'genres': x[0].metadata['genres'],
                "release_date": x[0].metadata['release_date'],
                "cast": x[0].metadata['cast'],
                "crew": x[0].metadata['crew'],
                "collection": x[0].metadata['collection'],
                "budget": x[0].metadata['budget'],
                "revenue": x[0].metadata['revenue'],
                "runtime": x[0].metadata['runtime'],
                "language": x[0].metadata['language'],
                "popularity": x[0].metadata['popularity'],
                "synopsis": x[0].metadata['synopsis'],
                "poster_path": x[0].metadata['poster_path'],
                "homepage": x[0].metadata['homepage'],
                'similarity_score': round(x[1], 2),
            }
            top_ten.append(movie_metadata)

        df_top_ten = pd.DataFrame(top_ten)
        return df_top_ten

    except Exception as e:
        print(f"Error during query: {e}")
        return None

import os

os.makedirs(name="data/final", exist_ok=True)
metadata.to_csv("data/final/final_metadata.csv")

class Recommender:
    def __init__(self, vectorstore, metadata_path):

        self.vectorstore = vectorstore
        self.metadata = pd.read_csv(metadata_path)

    @classmethod
    def from_weaviate(cls, metadata_path, embedding_model_args, weaviate_args ):
        client = weaviate.Client(
            url=weaviate_args['url'],
            auth_client_secret=weaviate.AuthApiKey(fetch_api_key(weaviate_args['ak_name'], False))
        )
        embeddings = HuggingFaceEmbeddings(
          model_name=embedding_model_args['model_name'],
          model_kwargs=embedding_model_args['model_kwargs']
        )
        vectorstore = Weaviate(client=client,embedding=embeddings,
                      index_name=weaviate_args['index_name'],
                      text_key=weaviate_args['text_key'],
                      by_text=weaviate_args['by_text'],
                     attributes=weaviate_args['attributes'],
                     )
        rec = cls(metadata_path = metadata_path, vectorstore=vectorstore)
        return rec

    def guess_movie(self, keyword):
        return self.metadata[self.metadata['title'].str.contains(keyword)]['title'].values[0]

    def get_recommendations_by_id(self, tmdb_id, k=10):
        if tmdb_id not in self.metadata.id:
            raise ValueError(f"Id '{tmdb_id}' not found in indices")
        querry = self.metadata[self.metadata['id'] == tmdb_id][['soup']].values[0]
        return self.recommend(querry[0], k)

    def get_recommendations_by_title(self, title, k=10):
        if title not in list(self.metadata.title):
            raise ValueError(f"title '{title}' not found in indices")
        querry = self.metadata[self.metadata['title'] == title][['soup']].values[0]
        return self.recommend(querry[0], k)

    def get_recommendations_by_keywords(self, keyword, k = 10):
        title = self.guess_movie(keyword)
        return self.get_recommendations_by_title(title, k = k)

    def recommend(self, query, k):
        try:
            results = self.vectorstore.similarity_search_with_score(query, k=k+1)

            top_k = []

            for x in results[1:]:

                movie_metadata = {
                    'movie': x[0].metadata['movie'],
                    "tmdb_id": x[0].metadata['tmdb_id'],
                    "imdb_id": x[0].metadata['imdb_id'],
                    'genres': x[0].metadata['genres'],
                    "release_date": x[0].metadata['release_date'],
                    "cast": x[0].metadata['cast'],
                    "crew": x[0].metadata['crew'],
                    "collection": x[0].metadata['collection'],
                    "budget": x[0].metadata['budget'],
                    "revenue": x[0].metadata['revenue'],
                    "runtime": x[0].metadata['runtime'],
                    "language": x[0].metadata['language'],
                    "popularity": x[0].metadata['popularity'],
                    "synopsis": x[0].metadata['synopsis'],
                    "poster_path": x[0].metadata['poster_path'],
                    "homepage": x[0].metadata['homepage'],
                    'similarity_score': round(x[1], 2),
                }
                top_k.append(movie_metadata)

            df_top_k = pd.DataFrame(top_k)
            return df_top_k

        except Exception as e:
            print(f"Error during query: {e}")
            return None


import yaml

def load_kwargs(fpath):
    with open(fpath, 'r') as file:
        config = yaml.safe_load(file)
    return config

rec = Recommender.from_weaviate(**load_kwargs("config/weaviate.yaml"))

rec.get_recommendations_by_keywords("Batman")

metadata.columns
